abstract cantor_basis test
