+++ src/f_log_mul.rs
@@ -90,3 +90,33 @@ pub fn walsh(data: &mut [Multiplier], size: usize) {
 }
 
 
+fn bitpoly_mul16(mut a: Wide, mut b: Wide) -> Wide {
+    let mut r: Wide =0;
+    for i in 0..FIELD_BITS {
+        if (b>>i) & 1 != 0 { r ^= (a<<i); }
+    }
+    r
+}
+
+fn gf_mul_bitpoly_reduced(mut a: Elt, mut b: Elt) -> Elt {
+    use core::convert::TryInto;
+    let len = FIELD_BITS;
+    let mut r: Wide = bitpoly_mul16(a as Wide,b as Wide);
+    let red : Wide = (1 << FIELD_BITS) + (GENERATOR as Wide);
+    for i in (len..=(len*2-1)).rev() {
+        if r & (1<<i) != 0 { r ^= (red<<(i-len)); }
+    }
+    r.try_into().unwrap()
+}
+
+#[test]
+fn cantor_basis() {
+    for w in BASE.windows(2) {
+        let b = w[1];
+        let square = gf_mul_bitpoly_reduced(b,b);
+        let a = w[0];
+        // let eq = if a == (square ^ b) { "==" } else { "!=" };
+        // println!("{:#b} {} {:#b}\n", a, eq, square ^ b);
+        assert_eq!(a, square ^ b);
+    }
+}
